+++ lib/cucumber/core/compiler.rb
@@ -1,11 +1,12 @@
 # frozen_string_literal: true
 
 require 'cucumber/core/test/case'
+require 'cucumber/core/test/data_table'
+require 'cucumber/core/test/doc_string'
+require 'cucumber/core/test/empty_multiline_argument'
+require 'cucumber/core/test/hook_step'
 require 'cucumber/core/test/step'
 require 'cucumber/core/test/tag'
-require 'cucumber/core/test/doc_string'
-require 'cucumber/core/test/data_table'
-require 'cucumber/core/test/empty_multiline_argument'
 require 'cucumber/messages'
 
 module Cucumber